rename sdks to Self and publish
