+++ sdk/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@selfxyz/core",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "repository": {
     "type": "git",
     "url": "https://github.com/celo-org/self"
